requirements.txt
openai>=1.12.0
spacy>=3.7.2
python-dotenv>=1.0.0
rich>=13.7.0
pyyaml>=6.0.1
tiktoken>=0.5.2
tenacity>=8.2.3
typing-extensions>=4.9.0
colorama>=0.4.6
requests>=2.31.0
aiohttp>=3.9.1
numpy>=1.24.3
tqdm>=4.66.1
pytest>=8.0.0
coverage>=7.4.1
mypy>=1.8.0
black>=24.1.1
flake8>=7.0.0 